🔨 refactor : alarmList item Component로 수정
